Make custom validators shared
