+++ src/app/shared/directives/birthday-validator.directive.spec.ts
@@ -1,0 +1,8 @@
+import { BirthdayValidatorDirective } from './birthday-validator.directive';
+
+describe('BirthdayValidatorDirective', () => {
+  it('should create an instance', () => {
+    const directive = new BirthdayValidatorDirective();
+    expect(directive).toBeTruthy();
+  });
+});
