+++ src/app/shared/directives/password-validator.directive.spec.ts
@@ -1,0 +1,8 @@
+import { PasswordValidatorDirective } from './password-validator.directive';
+
+describe('PasswordValidatorDirective', () => {
+  it('should create an instance', () => {
+    const directive = new PasswordValidatorDirective();
+    expect(directive).toBeTruthy();
+  });
+});
