+++ src/app/shared/shared.module.ts
@@ -2,6 +2,8 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MaterialModule } from './material/material.module';
 import { PasswordMatchValidatorDirective } from './directives/password-match-validator.directive';
+import { PasswordValidatorDirective } from './directives/password-validator.directive';
+import { BirthdayValidatorDirective } from './directives/birthday-validator.directive';
 
 
 @NgModule({
@@ -9,7 +11,9 @@ import { PasswordMatchValidatorDirective } from './directives/password-match-val
     CommonModule
   ],
   declarations: [
-    PasswordMatchValidatorDirective
+    PasswordMatchValidatorDirective,
+    PasswordValidatorDirective,
+    BirthdayValidatorDirective
   ],
   exports: [
     CommonModule,
